Made a very basic springy fella
The springy fella seems to rotate very quickly to the player, as soon as they detect them.

I believe this is because even though the rotation is tied to the velocity, which gradually increases, the direction of the vector is instantly set to the player. This causes the rotation to instantly snap to that new direction. No matter how small the magnitude.