Implemented Springy Fella's launch mechanic
As of now, the Springy Fella inherites from an IStandable interface. However, in the Player's Falling state, we directly check if the ground upon landing is a Springy Fella.

I have yet to figure out how to find a component that inherites from a certain class/interface, on a game object. If I knew how, this system might've been more expandable.

The Springy Fella has an OnStand(...) function, from the IStandable class. It basically checks whether the entity standing on it is a player, and gives the player a "guest state". In which the player is launched.