Updated the game code to match the new tools
All entities now have a dedicated class for movement settings, which can be expanded with inheritance. This eases the fluidity of creating and configuring new enemy types.

On top of that, a new function has been added to MovementBase.cs, which can be used for iterating an object to a halt, with drag instead of "grip". Maybe I'll use this in the Falling state for entities, but for now this is fine.